Added api version tracking and error reporting
If color_coded is updated (via vundle or other means), the native API
may need to be recompiled. This is determined by the VimL API and the
native API keeping track of their own versions; if the two versions do
not match, the native API should be recompiled and color_coded will not
work until this happens.